refactor: Extract checkHasCiphers and checkHasCozyOrg utils
This should make VaultUnlocker easier to test
